+++ Assets/ScalarTexture3DToSdfTexture3DProcedure/Example/Assets/ScalarTexture3DToSdfTexture3DProcedureExample.cs
@@ -9,25 +9,25 @@ using UnityEngine.Events;
 namespace Simplex.Procedures.Examples
 {
 	[ExecuteInEditMode]
-	public class MaskTexture3DToSdfTexture3DProcedureExample : MonoBehaviour
+	public class ScalarTexture3DToSdfTexture3DProcedureExample : MonoBehaviour
 	{
 		[SerializeField] Texture _sourceTexture3D = null;
 		[SerializeField] float _sourceValueThreshold = 0.5f;
-		[SerializeField] MaskTexture3DToSdfTexture3DProcedure.DownSampling _downSampling = MaskTexture3DToSdfTexture3DProcedure.DownSampling.None;
-		[SerializeField] MaskTexture3DToSdfTexture3DProcedure.Precision _precision = MaskTexture3DToSdfTexture3DProcedure.Precision._32;
+		[SerializeField] ScalarTexture3DToSdfTexture3DProcedure.DownSampling _downSampling = ScalarTexture3DToSdfTexture3DProcedure.DownSampling.None;
+		[SerializeField] ScalarTexture3DToSdfTexture3DProcedure.Precision _precision = ScalarTexture3DToSdfTexture3DProcedure.Precision._32;
 		[SerializeField] bool _addBorders = false;
 		[SerializeField] bool _showSource = false;
 
 		[Header("Output")]
 		[SerializeField] UnityEvent<RenderTexture> _sdfTexture3DEvent = new UnityEvent<RenderTexture>();
 
-		MaskTexture3DToSdfTexture3DProcedure _generator;
+		ScalarTexture3DToSdfTexture3DProcedure _generator;
 
 
 		void OnEnable()
 		{
 			_generator?.Release();
-			_generator = new MaskTexture3DToSdfTexture3DProcedure();
+			_generator = new ScalarTexture3DToSdfTexture3DProcedure();
 		}