+++ Assets/ScalarTexture3DToSdfTexture3DProcedure/Runtime/ScalarTexture3DToSdfTexture3DProcedure.cs
@@ -1,5 +1,5 @@
 /*
-	Copyright © Carl Emil Carlsen 2021-2024
+	Copyright © Carl Emil Carlsen 2024
 	http://cec.dk
 */
 
@@ -9,7 +9,7 @@ using UnityEngine.Rendering;
 
 namespace Simplex.Procedures
 {
-	public class MaskTextureToSdfTextureProcedure
+	public class ScalarTexture3DToSdfTexture3DProcedure
 	{
 		RenderTexture _sdfTexture;
 		RenderTexture _floodTexture;
@@ -20,12 +20,11 @@ namespace Simplex.Procedures
 		int _DistKernel;
 		int _ShowSeedsKernel;
 
-		LocalKeyword _ADD_BORDER;
+		LocalKeyword _ADD_BORDERS;
 
-		Vector2Int _groupThreadCount;
+		Vector3Int _groupThreadCount;
 
-		const int threadGroupWidth = 8; // Must match define in compute shader.
-		const string addBorderKeyword = "ADD_BORDER";
+		const int threadGroupLength = 8; // Must match define in compute shader.
 
 		public RenderTexture sdfTexture => _sdfTexture;
 
@@ -45,9 +44,9 @@ namespace Simplex.Procedures
 		}
 
 
-		public MaskTextureToSdfTextureProcedure()
+		public ScalarTexture3DToSdfTexture3DProcedure()
 		{
-			_computeShader = Object.Instantiate( Resources.Load<ComputeShader>( nameof( MaskTextureToSdfTextureProcedure ) ) );
+			_computeShader = Object.Instantiate( Resources.Load<ComputeShader>( nameof( ScalarTexture3DToSdfTexture3DProcedure ) ) );
 			_computeShader.hideFlags = HideFlags.HideAndDontSave;
 
 			_SeedKernel = _computeShader.FindKernel( nameof( _SeedKernel ) );
@@ -55,21 +54,23 @@ namespace Simplex.Procedures
 			_DistKernel = _computeShader.FindKernel( nameof( _DistKernel ) );
 			_ShowSeedsKernel = _computeShader.FindKernel( nameof( _ShowSeedsKernel ) );
 
-			_ADD_BORDER = new LocalKeyword( _computeShader, nameof( _ADD_BORDER ) );
+			_ADD_BORDERS = new LocalKeyword( _computeShader, nameof( _ADD_BORDERS ) );
 		}
 
 
 		public void Update
 		(
 			Texture sourceTexture, float sourceValueThreshold, 
-			DownSampling downSampling = DownSampling.None, Precision precision = Precision._32, bool addBorder = false,
+			DownSampling downSampling = DownSampling.None, Precision precision = Precision._32, bool addBorders = false,
 			bool _showSource = false
 		){
-
 			if( !sourceTexture ) return;
+			if( sourceTexture.dimension != TextureDimension.Tex3D ) throw new System.Exception( "sourceTexture must be a 3D texture." );
+
+			int resolutionZ = sourceTexture is RenderTexture ? ( sourceTexture as RenderTexture ).volumeDepth : ( sourceTexture as Texture3D ).depth;
 
 			// Ensure and adapt resources.
-			Vector2Int resolution = new Vector2Int( sourceTexture.width, sourceTexture.height );
+			Vector3Int resolution = new Vector3Int( sourceTexture.width, sourceTexture.height, resolutionZ );
 			switch( downSampling ) {
 				case DownSampling.Half: resolution /= 2; break;
 				case DownSampling.Quater: resolution /= 4; break;
@@ -80,52 +81,53 @@ namespace Simplex.Procedures
 				case Precision._16: sdfFormat = GraphicsFormat.R16_UNorm; break;	// ... and let the rest of the precision formats follow the same style.
 				default: sdfFormat = GraphicsFormat.R32_SFloat; break;				// ... R32 does not have a UNorm version, so we choose SFloat and use as unsigned.
 			}
-			if( !_sdfTexture || _sdfTexture.width != resolution.x || _sdfTexture.height != resolution.y || _sdfTexture.graphicsFormat != sdfFormat ) {
+			if( !_sdfTexture || _sdfTexture.width != resolution.x || _sdfTexture.height != resolution.y || _sdfTexture.volumeDepth != resolution.z || _sdfTexture.graphicsFormat != sdfFormat ) {
 				_sdfTexture?.Release();
-				_sdfTexture = CreateTexture( "SdfTexture", resolution, sdfFormat );
+				_sdfTexture = CreateTexture3D( "SdfTexture", resolution, sdfFormat );
 				_computeShader.SetTexture( _ShowSeedsKernel, ShaderIDs._SdfTex, _sdfTexture );
 				_computeShader.SetTexture( _DistKernel, ShaderIDs._SdfTex, _sdfTexture );
 			}
-			if( !_floodTexture || _floodTexture.width != resolution.x || _floodTexture.height != resolution.y ) {
+			if( !_floodTexture || _floodTexture.width != resolution.x || _floodTexture.height != resolution.y || _floodTexture.volumeDepth != resolution.z ) {
 				_floodTexture?.Release();
-				_floodTexture = CreateTexture( "FloodTexture", resolution, GraphicsFormat.R32G32B32A32_UInt );
+				_floodTexture = CreateTexture3D( "FloodTexture", resolution, GraphicsFormat.R32G32B32A32_UInt );
 				_computeShader.SetTexture( _SeedKernel, ShaderIDs._FloodTex, _floodTexture );
 				_computeShader.SetTexture( _FloodKernel, ShaderIDs._FloodTex, _floodTexture );
 				_computeShader.SetTexture( _DistKernel, ShaderIDs._FloodTexRead, _floodTexture );
 				_computeShader.SetTexture( _ShowSeedsKernel, ShaderIDs._FloodTexRead, _floodTexture );
-				_computeShader.SetInts( ShaderIDs._Resolution, new int[]{ resolution.x, resolution.y } );
-				_computeShader.SetVector( ShaderIDs._TexelSize, _sdfTexture.texelSize );
-				_groupThreadCount = new Vector2Int(
-					Mathf.CeilToInt( resolution.x / (float) threadGroupWidth ),
-					Mathf.CeilToInt( resolution.y / (float) threadGroupWidth )
+				_computeShader.SetInts( ShaderIDs._Resolution, new int[]{ resolution.x, resolution.y, resolution.z } );
+				_computeShader.SetVector( ShaderIDs._TexelSize, TexelSize3D( _sdfTexture ) );
+				_groupThreadCount = new Vector3Int(
+					Mathf.CeilToInt( resolution.x / (float) threadGroupLength ),
+					Mathf.CeilToInt( resolution.y / (float) threadGroupLength ),
+					Mathf.CeilToInt( resolution.z / (float) threadGroupLength )
 				);
 			}
 
 			// Set keywords.
-			if( _computeShader.IsKeywordEnabled( _ADD_BORDER ) != addBorder ) _computeShader.SetKeyword( _ADD_BORDER, addBorder );
+			if( _computeShader.IsKeywordEnabled( _ADD_BORDERS ) != addBorders ) _computeShader.SetKeyword( _ADD_BORDERS, addBorders );
 
 			// Seed.
 			_computeShader.SetTexture( _SeedKernel,  ShaderIDs._SeedTexRead, sourceTexture );
 			_computeShader.SetFloat( ShaderIDs._SeedThreshold, sourceValueThreshold );
-			_computeShader.Dispatch( _SeedKernel, _groupThreadCount.x, _groupThreadCount.y, 1 );
+			_computeShader.Dispatch( _SeedKernel, _groupThreadCount.x, _groupThreadCount.y, _groupThreadCount.z );
 
 			// Show seeds.
 			if( _showSource ) {
-				_computeShader.Dispatch( _ShowSeedsKernel, _groupThreadCount.x, _groupThreadCount.y, 1 );
+				_computeShader.Dispatch( _ShowSeedsKernel, _groupThreadCount.x, _groupThreadCount.y, _groupThreadCount.z );
 				return;
 			}
 
 			// Flood.
-			int sizeMax = Mathf.Max( resolution.x, resolution.y );
+			int sizeMax = Mathf.Max( Mathf.Max( resolution.x, resolution.y ), resolution.z );
 			int stepMax = (int) Mathf.Log( Mathf.NextPowerOfTwo( sizeMax ), 2 ); // 2^c_maxSteps is max image size on x and y
 			for( int n = stepMax; n >= 0; n-- ) {
 				int stepSize = n > 0 ? (int) Mathf.Pow( 2, n ) : 1;
 				_computeShader.SetInt( ShaderIDs._StepSize, stepSize );
-				_computeShader.Dispatch( _FloodKernel, resolution.x, resolution.y, 1 );
+				_computeShader.Dispatch( _FloodKernel, resolution.x, resolution.y, resolution.z );
 			}
 
 			// Compute SDF.
-			_computeShader.Dispatch( _DistKernel, _groupThreadCount.x, _groupThreadCount.y, 1 );
+			_computeShader.Dispatch( _DistKernel, _groupThreadCount.x, _groupThreadCount.y, _groupThreadCount.z );
 		}
 
 
@@ -138,10 +140,20 @@ namespace Simplex.Procedures
 		}
 
 
-		static RenderTexture CreateTexture( string name, Vector2Int resolution, GraphicsFormat format )
+		static Vector3 TexelSize3D( Texture t )
+		{
+			Vector3 texelSize = t.texelSize;
+			int resolutionZ = t is RenderTexture ? ( t as RenderTexture ).volumeDepth : ( t as Texture3D ).depth;
+			texelSize.z = 1f / resolutionZ;
+			return texelSize;
+		}
+
+		static RenderTexture CreateTexture3D( string name, Vector3Int resolution, GraphicsFormat format )
 		{
 			RenderTexture rt = new RenderTexture( resolution.x, resolution.y, 0, format, 0 );
 			rt.name = name;
+			rt.dimension = TextureDimension.Tex3D;
+			rt.volumeDepth = resolution.z;
 			rt.autoGenerateMips = false;
 			rt.enableRandomWrite = true;
 			rt.Create();