+++ Assets/ScalarTextureToSdfTextureProcedure/Example/Assets/ScalarTextureToSdfTextureProcedureExample.cs
@@ -9,25 +9,25 @@ using UnityEngine.Events;
 namespace Simplex.Procedures.Examples
 {
 	[ExecuteInEditMode]
-	public class MaskTextureToSdfTextureProcedureExample : MonoBehaviour
+	public class ScalarTextureToSdfTextureProcedureExample : MonoBehaviour
 	{
 		[SerializeField] Texture _sourceTexture = null;
 		[SerializeField] float _sourceValueThreshold = 0.5f;
-		[SerializeField] MaskTextureToSdfTextureProcedure.DownSampling _downSampling = MaskTextureToSdfTextureProcedure.DownSampling.None;
-		[SerializeField] MaskTextureToSdfTextureProcedure.Precision _precision = MaskTextureToSdfTextureProcedure.Precision._32;
+		[SerializeField] ScalarTextureToSdfTextureProcedure.DownSampling _downSampling = ScalarTextureToSdfTextureProcedure.DownSampling.None;
+		[SerializeField] ScalarTextureToSdfTextureProcedure.Precision _precision = ScalarTextureToSdfTextureProcedure.Precision._32;
 		[SerializeField] bool _addBorder = false;
 		[SerializeField] bool _showSource = false;
 
 		[Header("Output")]
 		[SerializeField] UnityEvent<RenderTexture> _sdfTextureEvent = new UnityEvent<RenderTexture>();
 
-		MaskTextureToSdfTextureProcedure _generator;
+		ScalarTextureToSdfTextureProcedure _generator;
 
 
 		void OnEnable()
 		{
 			_generator?.Release();
-			_generator = new MaskTextureToSdfTextureProcedure();
+			_generator = new ScalarTextureToSdfTextureProcedure();
 		}